+++ device-bootloaders/CHANGELOG.md
@@ -13,12 +13,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [SW-18602] Added support to rescan a partition after a FW image write.
 - [SW-15748] Added a separate DM events category for events directly reported to DM userspace, routed trace buffer full event to DM.
 - [SW-18460] Adding voltage verification through PMIC and retries for setting the voltage.
+- [SW-18145] Reset the I2C module if it hangs.
+  - Added new APIs for enable, disable and abort of I2C interface.
+  - Added a workaround to reset the I2C module if it hangs.
 ### Changed
 - [SW-18208] Enabling PMIC Firmware update.
 - [SW-18499] Masking out bad Minion Shire PVT sampling.
 -  Improving the logging of PMIC FW update.
 ### Deprecated
 ### Removed
+- Workaround delays in I2C APIs.
 ### Fixed
 - Fix bug in getting the inactive PMIC slot info
 - [SW-18720] Fix bogus value in SP trace header type